Refactored HTML and CSS
Added Responsiveness down to 320px (The minimum screen size for cellphone screens). Changed class names based on a new naming convention I'm still figuring out, but that is way more semantic than that of the original code. The added responsiveness required an addition of a new flex-container inside card__product, as well as changing the behaviour of card__product with media queries when the screen_width is smaller than 380px.